Added checks for invalid entries for the headingsGsWcaActivity.java

+++ src/team/pfm/com/Calculation.java
@@ -32,11 +32,11 @@ public class Calculation {
 	//Calculates head wind and cross wind given wind speed wind direction and runway direction
 	// returns 4 values as an array that denote speed and direction of winds
 	//0deg is due north, 180deg is due south, 90deg is east, 270deg is due west
-	public static double[] calcRunWinds(double windSpe,int windDir, int runwayDir){
+	public static double[] calcRunWinds(double windSpe,double windDir, double runwayDir){
 		double xwindDir;
 		double hwindDir;
 		double []val = new double[4];
-		int angle = Math.abs(windDir - runwayDir);
+		double angle = Math.abs(windDir - runwayDir);
 		//determines what direction the xwind is coming from
 		if(windDir> runwayDir && windDir<(runwayDir + 180) ){
 			xwindDir = 1;//VALUE 1 DENOTES XWIND COMING FROM THE RIGHT
@@ -61,7 +61,7 @@ public class Calculation {
 	
 	 //calculates wind speed given ground speed, true air speed, course and heading
 	//the calculation was configured using a system of equations 
-	public static double calcWindSpe(double groSpe, double trueAir, int course , int heading){
+	public static double calcWindSpe(double groSpe, double trueAir, double course , double heading){
 		double wca = heading - course;
 		double windSpe= Math.sqrt( Math.pow(groSpe, 2) + Math.pow(trueAir,2) -(2*groSpe* trueAir * Math.cos(Math.toRadians(wca))));
 		return windSpe;
@@ -70,9 +70,9 @@ public class Calculation {
 	//calculates wind directions given ground speed, true air speed, course and heading
 	//the calculation was configured using a system of equations 
 	//all data is input as radians
-	public static double calcWindDir(double groSpe, double trueAir, int course , int heading){
+	public static double calcWindDir(double groSpe, double trueAir, double course , double heading){
 		double ws = calcWindSpe(groSpe, trueAir, course, heading);
-		int wca = heading - course;
+		double wca = heading - course;
 		
 		return heading + Math.toDegrees(Math.asin((groSpe/ws) * Math.sin(Math.toRadians(wca)) ));
 		
@@ -80,7 +80,7 @@ public class Calculation {
 	}
 	
 	//calculates the wind correction angle given wind speed, wind direction(from source), airspeed, and course
-	public static double calcCorAng(double windSpe, int windDir, double trueAir, int course){
+	public static double calcCorAng(double windSpe, double windDir, double trueAir, double course){
 		double windDirDownwind = windDir + 180; //getting the downwind direction of the wind
 		if(windDirDownwind>=360){
 			windDirDownwind= windDirDownwind - 360;
@@ -92,7 +92,7 @@ public class Calculation {
 	}
 	
 	//calculates the ground speed given wind speed, wind direction, air speed, and course
-	public static double calcGroSpeed(double windSpe, int windDir,int course, double trueAir){
+	public static double calcGroSpeed(double windSpe, double windDir,double course, double trueAir){
 		double dwdir = windDir + 180;
 		if(dwdir>=360){
 			dwdir= dwdir - 360;
@@ -103,7 +103,7 @@ public class Calculation {
 	}
 	
 	//calculates the heading given wind speed, wind direction, air speed, and course
-	public static double calcHead(double windSpe, int windDir, int course, double trueAir){
+	public static double calcHead(double windSpe, double windDir, double course, double trueAir){
 		double wca = calcCorAng(windSpe, windDir, trueAir, course);
 		return wca + course;
 	}
